pyproject.toml
[project]
name = "agentllm"
version = "0.1.0"
description = "Custom LiteLLM provider for Agno agents - OpenAI-compatible API for Agno"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.2.8",
    "jira>=3.0.0",
    "litellm[proxy]>=1.79.1",
    "loguru>=0.7.0",
    "nox>=2025.10.16",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "sqlalchemy>=2.0.0",
    "google-genai>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "html-to-markdown>=1.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
